+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.2.0
+
+BREAKING CHANGES:
+
+* Provider has been renamed to `kerberos`
+* resource/keytab\_file: Renamed to `kerberos_keytab`
+
+FEATURES:
+
+* New Data Source: `kerberos_token`
+
 ## 0.1.0
 
 FEATURES: